pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ansify"
version = "0.1.0"
description = "Awesome `ansify` is a Python CLI to create ANSI/ASCII art from images."
readme = "README.md"
authors = [
  "lonsty <lonsty@sina.com>"
]
license = "MIT"
repository = "https://github.com/lonsty/ansify"
homepage = "https://github.com/lonsty/ansify"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  # Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  # Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"ansify" = "ansify.__main__:app"

[tool.poetry.dependencies]
python = "^3.6.9"
importlib_metadata = {version = "^1.6.0", python = "<3.8"}
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^10.0.0"
Pillow = "^8.1.2"
numpy = "^1.19.5"
requests = "^2.25.1"
urwid = "^2.1.2"

[tool.poetry.dev-dependencies]
darglint = "^1.7.0"
isort = "^5.8.0"
pyupgrade = "^2.11.0"
black = "^20.8b1"
mypy = "^0.790"
bandit = "^1.7.0"
safety = "^1.10.3"
pytest = "^6.2.2"
pylint = "^2.7.2"
pydocstyle = "^5.1.1"
pre-commit = "^2.11.1"

[tool.black]
# https://github.com/psf/black
line-length = 120
target-version = ["py36"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 120
